+++ src/include/getVarDegree.h
@@ -2,6 +2,6 @@
 #define GETVARDEGREE_H
 #include <vector>
 
-std::vector<int> getVarDegree(int dim, int n);
+std::vector<int> getVarDegree(size_t dim, size_t n);
 
 #endif
